requirements.txt
pycrypto==2.6
pyasn1==0.1.7
Twisted==13.1.0
wsgiref==0.1.2
zope.interface==4.0.5